Attempt to improve triggers
The libvirt nvme disk create/destroy code is now all
in trigger blocks.

However, in the 'make' test, the files are recreated after deleting
them. This is some bug in vagrant-libvirt, I believe.
The impact is that there is a 197KB sparse file recreated for each
nvme disk AFTER deleting the VMs.